+++ src/main/java/board/Board.java
@@ -1,0 +1,14 @@
+package board;
+
+import frame.Frames;
+
+public class Board {
+
+    private final Name name;
+    private final Frames frames;
+
+    public Board(Name name, Frames frames) {
+        this.name = name;
+        this.frames = frames;
+    }
+}
