+++ src/main/java/board/Name.java
@@ -1,3 +1,5 @@
+package board;
+
 import java.util.Objects;
 
 public class Name {
