+++ src/main/java/frame/Frame.java
@@ -1,4 +1,9 @@
 package frame;
 
+import score.ScoreInfo;
+
+import java.util.List;
+
 public interface Frame {
+    NormalFrame nextFrame(List<ScoreInfo> scores);
 }
