+++ src/main/java/frame/Frames.java
@@ -1,0 +1,12 @@
+package frame;
+
+import java.util.List;
+
+public class Frames {
+
+    private final List<Frame> frames;
+
+    public Frames(List<Frame> frames) {
+        this.frames = frames;
+    }
+}
