+++ src/main/java/frame/LastFrame.java
@@ -1,4 +1,12 @@
 package frame;
 
+import score.ScoreInfo;
+
+import java.util.List;
+
 public class LastFrame implements Frame {
+    @Override
+    public NormalFrame nextFrame(List<ScoreInfo> scores) {
+        return null;
+    }
 }
