+++ src/test/java/NameTest.java
@@ -1,3 +1,4 @@
+import board.Name;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
